Refactor: AuthTopBar와 PageHeader를 ScreenHeader 셸로 흡수한다
