+++ components/templates/ForgotPasswordTemplate.tsx
@@ -1,5 +1,5 @@
 import { ForgotPasswordForm } from "@/components/organisms";
-import { AuthTopBar } from "@/components/molecules/AuthTopBar";
+import { AuthTopBar } from "@/components/molecules";
 import { DailyLoopAuthTemplate } from "@/components/templates/DailyLoopAuthTemplate";
 import { ROUTES } from "@/config/routes";
