+++ components/templates/ResetPasswordTemplate.tsx
@@ -1,5 +1,5 @@
 import { ResetPasswordForm } from "@/components/organisms";
-import { AuthTopBar } from "@/components/molecules/AuthTopBar";
+import { AuthTopBar } from "@/components/molecules";
 import { DailyLoopAuthTemplate } from "@/components/templates/DailyLoopAuthTemplate";
 import { ROUTES } from "@/config/routes";
